requirements.txt
streamlit
openai
python-dotenv
streamlit-extras
streamlit-elements
graphviz
pillow
scikit-learn 